Abstracted out testing code
