+++ pkg/controllers/dashboard_test.go
@@ -17,17 +17,10 @@ limitations under the License.
 package controllers
 
 import (
-	"reflect"
 	"testing"
-	"time"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/util/diff"
 	k8sfake "k8s.io/client-go/kubernetes/fake"
-	core "k8s.io/client-go/testing"
-	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/record"
 
 	grafanacontroller "kubernetes-grafana-controller/pkg/apis/grafana/v1alpha1"
@@ -35,45 +28,6 @@ import (
 	informers "kubernetes-grafana-controller/pkg/client/informers/externalversions"
 	grafana "kubernetes-grafana-controller/pkg/grafana/fake"
 )
-
-var (
-	alwaysReady        = func() bool { return true }
-	noResyncPeriodFunc = func() time.Duration { return 0 }
-)
-
-const (
-	FAKE_UID = "fakeUID"
-)
-
-type fixture struct {
-	t *testing.T
-
-	client        *fake.Clientset
-	kubeclient    *k8sfake.Clientset
-	grafanaClient *grafana.ClientFake
-
-	// Objects to put in the store.
-	grafanaDashboardLister []*grafanacontroller.GrafanaDashboard
-
-	// Actions expected to happen on the client.
-	kubeactions       []core.Action
-	actions           []core.Action
-	grafanaPostedJson *string
-
-	// Objects from here preloaded into NewSimpleFake.
-	kubeobjects []runtime.Object
-	objects     []runtime.Object
-}
-
-func newFixture(t *testing.T) *fixture {
-	f := &fixture{}
-	f.t = t
-	f.objects = []runtime.Object{}
-	f.kubeobjects = []runtime.Object{}
-	f.grafanaPostedJson = nil
-
-	return f
-}
 
 func newGrafanaDashboard(name string, dashboardJson string) *grafanacontroller.GrafanaDashboard {
 	return &grafanacontroller.GrafanaDashboard{
@@ -88,7 +42,7 @@ func newGrafanaDashboard(name string, dashboardJson string) *grafanacontroller.G
 	}
 }
 
-func (f *fixture) newController() (*Controller, informers.SharedInformerFactory) {
+func newDashboardController(f *fixture) (*Controller, informers.SharedInformerFactory) {
 	f.client = fake.NewSimpleClientset(f.objects...)
 	f.kubeclient = k8sfake.NewSimpleClientset(f.kubeobjects...)
 	f.grafanaClient = grafana.NewGrafanaClientFake("https://example.com", FAKE_UID)
@@ -108,161 +62,20 @@ func (f *fixture) newController() (*Controller, informers.SharedInformerFactory)
 	return c, i
 }
 
-func (f *fixture) run(grafanaDashboardName string) {
-	f.runController(grafanaDashboardName, true, false)
-}
-
-func (f *fixture) runExpectError(grafanaDashboardName string) {
-	f.runController(grafanaDashboardName, true, true)
-}
-
-func (f *fixture) runController(grafanaDashboardName string, startInformers bool, expectError bool) {
-	c, i := f.newController()
-	if startInformers {
-		stopCh := make(chan struct{})
-		defer close(stopCh)
-		i.Start(stopCh)
-	}
-
-	err := c.syncer.(*DashboardSyncer).syncHandler(NewWorkQueueItem(grafanaDashboardName, Dashboard, ""))
-	if !expectError && err != nil {
-		f.t.Errorf("error syncing dashboard: %v", err)
-	} else if expectError && err == nil {
-		f.t.Error("expected error syncing dashboard, got nil")
-	}
-
-	actions := filterInformerActions(f.client.Actions())
-	for i, action := range actions {
-		if len(f.actions) < i+1 {
-			f.t.Errorf("%d unexpected actions: %+v", len(actions)-len(f.actions), actions[i:])
-			break
-		}
-
-		expectedAction := f.actions[i]
-		checkAction(expectedAction, action, f.t)
-	}
-
-	if len(f.actions) > len(actions) {
-		f.t.Errorf("%d additional expected actions:%+v", len(f.actions)-len(actions), f.actions[len(actions):])
-	}
-
-	k8sActions := filterInformerActions(f.kubeclient.Actions())
-	for i, action := range k8sActions {
-		if len(f.kubeactions) < i+1 {
-			f.t.Errorf("%d unexpected actions: %+v", len(k8sActions)-len(f.kubeactions), k8sActions[i:])
-			break
-		}
-
-		expectedAction := f.kubeactions[i]
-		checkAction(expectedAction, action, f.t)
-	}
-
-	if len(f.kubeactions) > len(k8sActions) {
-		f.t.Errorf("%d additional expected actions:%+v", len(f.kubeactions)-len(k8sActions), f.kubeactions[len(k8sActions):])
-	}
-
-	// test grafana client "actions"
-	if f.grafanaPostedJson != nil &&
-		*f.grafanaPostedJson != *f.grafanaClient.PostedJson {
-
-		f.t.Errorf("Expected posted dashboard json %s but found %s", *f.grafanaPostedJson, *f.grafanaClient.PostedJson)
-	}
-}
-
-// checkAction verifies that expected and actual actions are equal and both have
-// same attached resources
-func checkAction(expected, actual core.Action, t *testing.T) {
-	if !(expected.Matches(actual.GetVerb(), actual.GetResource().Resource) && actual.GetSubresource() == expected.GetSubresource()) {
-		t.Errorf("Expected\n\t%#v\ngot\n\t%#v", expected, actual)
-		return
-	}
-
-	if reflect.TypeOf(actual) != reflect.TypeOf(expected) {
-		t.Errorf("Action has wrong type. Expected: %t. Got: %t", expected, actual)
-		return
-	}
-
-	switch a := actual.(type) {
-	case core.CreateAction:
-		e, _ := expected.(core.CreateAction)
-		expObject := e.GetObject()
-		object := a.GetObject()
-
-		if !reflect.DeepEqual(expObject, object) {
-			t.Errorf("Action %s %s has wrong object\nDiff:\n %s",
-				a.GetVerb(), a.GetResource().Resource, diff.ObjectGoPrintDiff(expObject, object))
-		}
-	case core.UpdateAction:
-		e, _ := expected.(core.UpdateAction)
-		expObject := e.GetObject()
-		object := a.GetObject()
-
-		if !reflect.DeepEqual(expObject, object) {
-			t.Errorf("Action %s %s has wrong object\nDiff:\n %s",
-				a.GetVerb(), a.GetResource().Resource, diff.ObjectGoPrintDiff(expObject, object))
-		}
-	case core.PatchAction:
-		e, _ := expected.(core.PatchAction)
-		expPatch := e.GetPatch()
-		patch := a.GetPatch()
-
-		if !reflect.DeepEqual(expPatch, patch) {
-			t.Errorf("Action %s %s has wrong patch\nDiff:\n %s",
-				a.GetVerb(), a.GetResource().Resource, diff.ObjectGoPrintDiff(expPatch, patch))
-		}
-	}
-}
-
-// filterInformerActions filters list and watch actions for testing resources.
-// Since list and watch don't change resource state we can filter it to lower
-// noise level in our tests.
-func filterInformerActions(actions []core.Action) []core.Action {
-	ret := []core.Action{}
-	for _, action := range actions {
-		if len(action.GetNamespace()) == 0 &&
-			(action.Matches("list", "grafanadashboards") ||
-				action.Matches("watch", "grafanadashboards")) {
-			continue
-		}
-		ret = append(ret, action)
-	}
-
-	return ret
-}
-
-func (f *fixture) expectUpdateGrafanaUid(grafanaDashboard *grafanacontroller.GrafanaDashboard) {
-	action := core.NewUpdateAction(schema.GroupVersionResource{Resource: "grafanadashboards"}, grafanaDashboard.Namespace, grafanaDashboard)
-	// TODO: Until #38113 is merged, we can't use Subresource
-	//action.Subresource = "status"
-	f.actions = append(f.actions, action)
-}
-
-func (f *fixture) expectGrafanaDashboardPost(dashboardJson string) {
-	f.grafanaPostedJson = &dashboardJson
-}
-
-func getKey(grafanaDashboard *grafanacontroller.GrafanaDashboard, t *testing.T) string {
-	key, err := cache.DeletionHandlingMetaNamespaceKeyFunc(grafanaDashboard)
-	if err != nil {
-		t.Errorf("Unexpected error getting key for dashboard %v: %v", grafanaDashboard.Name, err)
-		return ""
-	}
-	return key
-}
-
 func TestCreatesGrafanaDashboard(t *testing.T) {
 
 	f := newFixture(t)
 	dashboardJson := "{ 'test': 'test' }"
 
 	dashboard := newGrafanaDashboard("test", dashboardJson)
+	item := NewWorkQueueItem(getKey(dashboard, t), Dashboard, "")
 
 	f.grafanaDashboardLister = append(f.grafanaDashboardLister, dashboard)
 	f.objects = append(f.objects, dashboard)
 
 	dashboard.Status.GrafanaUID = FAKE_UID
-	f.expectUpdateGrafanaUid(dashboard)
-	f.expectGrafanaDashboardPost(dashboardJson)
+	f.expectUpdateGrafanaObject(dashboard, dashboard.Namespace, "grafanadashboards")
+	f.expectGrafanaPost(dashboardJson)
 
-	f.run(getKey(dashboard, t))
+	f.runController(newDashboardController, item, true, false)
 }
